Repair reflector dataset and procedures
DRAGON5 can now execute this reflector dataset. The main change that
has been required is the use of the 15x15 geometry, copied from the
fuel. DRAGON5 revision 862 was originally used, but this geometry
update involves using the same revision as for the fuel (DRAGON5
revision 1487). Two procedures (CreaStepList and GetTechData) were
also adopted (copied) from the fuel modelling. The same JEFF-3.1.1
Draglib as for the fuel is used, i.e. the one produced with
https://github.com/IRSN/PyNjoy2016