Adapt procedures to enable fuel
The main changes concern the parameters of datasets (EvoNominale and
MakeBib1BU), which now use input/output objects. These objects are
still in XSM files for the moment, although linked lists could be
used as it may enable a speedup. Also, the history effect feature has
been removed as it is not used.