Correct an error on the O16 reflector composition
Each layer of a reflector is defined as a mixture of several
materials, including water and zircalloy. Oxygen 16 is present:
* in large quantities in water (H2O) and
* as traces in zircalloy.
This led a user to declare the isotope O16 twice, in that order.
DRAGON allows, as a feature, to declare the same isotope a second
time, but to overwrite a previous declaration. The oxygen 16 of the
water in the reflector was therefore ignored. Now the sum of both is
used. On Tihange startup, this long-standing error causes :
* an overestimation of the power at the center of 5 to 6%,
* an underestimation of the power on the edge of 5 to 7%,
* an underestimation of the reactivity of about 120 pcm.