Allow randomly sampled isotopes from TENDL
The reference results (JEFF-3.1.1) are very slightly modified (0.2
pcm on reactivity, ~0.1% on power). This is caused by the separation
in two of the calls to LIB: in CreaMix. The combinations of MIX
(keyword COMB) are moved to a second call, in order to be able to
insert possible modifications from TENDL.