Redesign reflector calc: separate geometry + compo
The geometry and composition of the reflector are separated from the
rest of the reflector procedure, to allow exporting them to Serpent.
Currently, the new ReflGeoCompo procedure still produces a Macrolib:
it will have to be a library object to reach the export goal.

As a side benefit, this separation clarifies the reflector calculation
procedure.