Add D2S -- auto convert from Dragon to Serpent
This conversion is as generic as possible, but is currently limited to
eighths symmetrical PWR assemblies. It can be used with any assembly
size, with any number and position of cells, with any compositions and
temperatures. Relying heavily on PyGan, it directly reads Dragon's
internal structure format (LCM objects) that are usually communicated
to its deterministic solvers. By design, the produced Serpent inputs
are friendly for human reading.

In the provided example, D2S is only executed on one single specific
assembly, but nothing should hinder an application on other cases.

Also, to feed D2S, Dragon's geometries and compositions are exported,
for the case of Tihange HZP startup.