Extend Draglib plots to any isotope and reaction
It is no longer necessary to specify the reactions to be plotted: if a
reaction is possible, it is plotted. This makes it possible to plot the
cross sections of any isotope. Threshold reactions are the subject of
particular attention, since each random sample does not necessarily have
the same number of points for plotting.

If Autolibs are available ('BIN-' + reaction), their content is now
automatically plotted.