Add Drakkar std-dev/skew/kurt standard errors plot
Skewness and kurtosis are now plotted on the Drakkar summary plot. On
this same plot, standard errors, due to the necessarily finite number of
nuclear data samples, are shown for :
* relative standard deviations,
* skewness and
* kurtosis.

Finally, a new plot (StandardError.py) is added to show the changes in
these standard errors as a function of the number of samples.