Correct degrees of freedom in std-dev estimators
Until now, we estimated the standard deviation by considering n degrees
of freedom, where n is the number of samples (ddof = 0, in Python
syntax). But, rigorously speaking, one degree of freedom is lost when
estimating the mean: we have to use (n-1) degrees of freedom instead
(ddof = 1, in Python syntax). This commit implements that.

Although previously erroneous, again rigorously speaking, this
correction is expected to have a very small impact, given the large
number of samples that are typically used.