Add detector response in Drakkar
The responses of the mobile fission chambers, particularly useful for
Fessenheim-1/2 and Bugey-2, is added with this commit, through calls to
IDET module.

Previously, the geometry was discretized into 2x2 by assembly using the
keywords SPLITX and SPLITY (automatic splitting). However, these
keywords are not compatible with the use of the IDET module
(non-programmed functionality). This has multiple, cascading
consequences.

Firstly, a 1x1 mesh being insufficient, all core geometries had to be
rewritten in 2x2 meshes by assembly, explicitly. Following that :
(1) The files of geometric descriptions have henceforth too many
    columns, 73 characters being the maximum imposed by CLE-2000
    language. An automatic carriage return is imposed in the .access
    scripts, at execution time. This makes it possible to keep the
    complete geometry in the description files, without any linebreak.
(2) Printouts of the THM module had to be deactivated, since it was not
    designed for explicit 2x2 meshes. This modification in Donjon is
    being reported for its future revisions.

This last point has, in turn, a consequence : non-regression was so far
achieved by recovering these power distribution outputs given by THM.
The non-regression system has therefore been rewritten. It is now based
on output objects, rather than output listing.

Of course, changing the non-regression system, accompanied by the
above-mentioned broad and inseparable modifications, involves a great
risk of regression. Extensive comparisons were therefore made. They
showed that effective multiplication factor was modified by 0.4 pcm and
that the power distribution was unaffected, for all Tihange-1 cases. A
small difference such as this one is expected, since axial
discretization has been redone : at the bottom of the core, a single
(unsplitted) mesh is now used for the layer without control rods,
instead of a single mesh splitted in two.