Add bayesian adjustments with tmp CASMO5 detector
The probability densities of detector activities, for each assemblies,
are now plotted. Bayesian weights are computed for the first time, based
on BMC and BFMC techniques.

An exaggerated checkerboard pattern appears with the Drakkar detectors,
as in the work of Frohlicher. The power distribution of Drakkar is thus
combined, temporarily, with the detector/power ratio computed by
CASMO5.