Remove SPH equivalence for detector cross sections
For a reason not yet understood, the SPH equivalence causes checkerboard
errors (related to the loading of Pyrex assemblies in checkerboard) on
the response of the detectors. Without it, this checkerboard error
vanishes. The SPH equivalence is therefore deactivated for the cross
sections of the detector isotope, i.e. the uranium 235 traces in the
central guide tube.

The SPH equivalence remains for all other cross sections, i.e. flux and
power distributions are not affected by this modification.