Add case used for PyNjoy's Draglib non-regression
This simple Rowland depleting pincell case is extracted from Dragon5
non-regression datasets. It has been used to check the non-regression
between Draglibs produced by PyNjoy, PyNjoy2012 and PyNjoy2016
(respectively based on NJOY99, NJOY2012 and NJOY2016).

The demonstration of non-regression cannot be rerun from the information
provided here, due to NJOY99 and 2012 closeness. This file is provided
rather for the record.