Acknowledge bashisms as such (for Debian/Ubuntu)
A path instead of a command is given to "command". This is not a problem
if Bash is used as default shell, such as Red Hat-based systems:
https://github.com/IRSN/SalinoPhD/blob/03a35dddd49ebdc06c122a5439524a22e6b5b1d0/.circleci/config.yml#L7
However, it is a bashism and causes problems, for example, on Debian or
Ubuntu systems (using Dash as default shell). This commit fixes this,
recognizing it as a bashism.

Moreover, all shell scripts have been tested against ShellCheck:
https://github.com/koalaman/shellcheck
No major issues were identified.